fix: ROUND(double precision, int) SQL 类型错误
- schema.sql: success_rate FLOAT→NUMERIC(6,4), avg_match_score FLOAT→NUMERIC(10,4)
- DB: DROP VIEW → ALTER TABLE → CREATE VIEW
- analyze_learning_data.py: SUM 计算加 ::NUMERIC 类型转换

+++ openclaw-workspaces/ai-order/scripts/analyze_learning_data.py
@@ -145,7 +145,7 @@ def analyze_threshold_tuning():
               COUNT(*) as total,
               SUM(CASE WHEN auto_matched THEN 1 ELSE 0 END) as confirmed,
               SUM(CASE WHEN NOT auto_matched THEN 1 ELSE 0 END) as corrected,
-              ROUND(SUM(CASE WHEN NOT auto_matched THEN 1 ELSE 0 END)::FLOAT / NULLIF(COUNT(*), 0) * 100, 2) as correction_rate
+              ROUND((SUM(CASE WHEN NOT auto_matched THEN 1 ELSE 0 END)::NUMERIC / NULLIF(COUNT(*), 0) * 100)::NUMERIC, 2) as correction_rate
             FROM order_corrections
             WHERE created_at >= CURRENT_DATE - INTERVAL '%s days'
             GROUP BY match_layer
